Refactor FraudNet namespace, isn't related specially to PUI
