+++ modules/ppcp-wc-gateway/src/FraudNet/FraudNetSourceWebsiteId.php
@@ -7,7 +7,7 @@
 
 declare(strict_types=1);
 
-namespace WooCommerce\PayPalCommerce\WcGateway\Gateway\PayUponInvoice;
+namespace WooCommerce\PayPalCommerce\WcGateway\FraudNet;
 
 /**
  * Class FraudNetSourceWebsiteId.
