+++ modules/ppcp-wc-gateway/src/Gateway/PayUponInvoice/PayUponInvoice.php
@@ -17,6 +17,7 @@ use WooCommerce\PayPalCommerce\ApiClient\Factory\CaptureFactory;
 use WooCommerce\PayPalCommerce\Button\Exception\RuntimeException;
 use WooCommerce\PayPalCommerce\Onboarding\Environment;
 use WooCommerce\PayPalCommerce\Session\SessionHandler;
+use WooCommerce\PayPalCommerce\WcGateway\FraudNet\FraudNet;
 use WooCommerce\PayPalCommerce\WcGateway\Gateway\PayPalGateway;
 use WooCommerce\PayPalCommerce\WcGateway\Helper\CheckoutHelper;
 use WooCommerce\PayPalCommerce\WcGateway\Helper\PayUponInvoiceHelper;
@@ -33,20 +34,6 @@ use WP_Error;
 class PayUponInvoice {
 
 	use TransactionIdHandlingTrait;
-
-	/**
-	 * The module URL.
-	 *
-	 * @var string
-	 */
-	protected $module_url;
-
-	/**
-	 * The FraudNet entity.
-	 *
-	 * @var FraudNet
-	 */
-	protected $fraud_net;
 
 	/**
 	 * The pui order endpoint.
@@ -70,20 +57,6 @@ class PayUponInvoice {
 	protected $settings;
 
 	/**
-	 * The environment.
-	 *
-	 * @var Environment
-	 */
-	protected $environment;
-
-	/**
-	 * The asset version.
-	 *
-	 * @var string
-	 */
-	protected $asset_version;
-
-	/**
 	 * The PUI helper.
 	 *
 	 * @var PayUponInvoiceHelper
@@ -96,13 +69,6 @@ class PayUponInvoice {
 	 * @var State
 	 */
 	protected $state;
-
-	/**
-	 * Whether the current page is the PPCP settings page.
-	 *
-	 * @var bool
-	 */
-	protected $is_ppcp_settings_page;
 
 	/**
 	 * Current PayPal settings page id.
@@ -133,63 +99,38 @@ class PayUponInvoice {
 	protected $capture_factory;
 
 	/**
-	 * The session handler
-	 *
-	 * @var SessionHandler
-	 */
-	protected $session_handler;
-
-	/**
 	 * PayUponInvoice constructor.
 	 *
-	 * @param string                      $module_url The module URL.
-	 * @param FraudNet                    $fraud_net The FraudNet entity.
 	 * @param PayUponInvoiceOrderEndpoint $pui_order_endpoint The PUI order endpoint.
 	 * @param LoggerInterface             $logger The logger.
 	 * @param Settings                    $settings The settings.
-	 * @param Environment                 $environment The environment.
-	 * @param string                      $asset_version The asset version.
 	 * @param State                       $state The onboarding state.
-	 * @param bool                        $is_ppcp_settings_page Whether page is PayPal settings poge.
 	 * @param string                      $current_ppcp_settings_page_id Current PayPal settings page id.
 	 * @param PayUponInvoiceProductStatus $pui_product_status The PUI product status.
 	 * @param PayUponInvoiceHelper        $pui_helper The PUI helper.
 	 * @param CheckoutHelper              $checkout_helper The checkout helper.
 	 * @param CaptureFactory              $capture_factory The capture factory.
-	 * @param SessionHandler              $session_handler The session handler.
 	 */
 	public function __construct(
-		string $module_url,
-		FraudNet $fraud_net,
 		PayUponInvoiceOrderEndpoint $pui_order_endpoint,
 		LoggerInterface $logger,
 		Settings $settings,
-		Environment $environment,
-		string $asset_version,
 		State $state,
-		bool $is_ppcp_settings_page,
 		string $current_ppcp_settings_page_id,
 		PayUponInvoiceProductStatus $pui_product_status,
 		PayUponInvoiceHelper $pui_helper,
 		CheckoutHelper $checkout_helper,
-		CaptureFactory $capture_factory,
-		SessionHandler $session_handler
+		CaptureFactory $capture_factory
 	) {
-		$this->module_url                    = $module_url;
-		$this->fraud_net                     = $fraud_net;
 		$this->pui_order_endpoint            = $pui_order_endpoint;
 		$this->logger                        = $logger;
 		$this->settings                      = $settings;
-		$this->environment                   = $environment;
-		$this->asset_version                 = $asset_version;
 		$this->state                         = $state;
-		$this->is_ppcp_settings_page         = $is_ppcp_settings_page;
 		$this->current_ppcp_settings_page_id = $current_ppcp_settings_page_id;
 		$this->pui_product_status            = $pui_product_status;
 		$this->pui_helper                    = $pui_helper;
 		$this->checkout_helper               = $checkout_helper;
 		$this->capture_factory               = $capture_factory;
-		$this->session_handler               = $session_handler;
 	}
 
 	/**
@@ -234,11 +175,6 @@ class PayUponInvoice {
 
 				return $data;
 			}
-		);
-
-		add_action(
-			'wp_enqueue_scripts',
-			array( $this, 'register_assets' )
 		);
 
 		add_action(
@@ -586,32 +522,5 @@ class PayUponInvoice {
 				}
 			}
 		);
-	}
-
-	/**
-	 * Registers PUI assets.
-	 */
-	public function register_assets(): void {
-		$gateway_settings = get_option( 'woocommerce_ppcp-pay-upon-invoice-gateway_settings' );
-		$gateway_enabled  = $gateway_settings['enabled'] ?? '';
-		if ( $gateway_enabled === 'yes' && ! $this->session_handler->order() && ( is_checkout() || is_checkout_pay_page() ) ) {
-			wp_enqueue_script(
-				'ppcp-pay-upon-invoice',
-				trailingslashit( $this->module_url ) . 'assets/js/pay-upon-invoice.js',
-				array(),
-				$this->asset_version,
-				true
-			);
-
-			wp_localize_script(
-				'ppcp-pay-upon-invoice',
-				'FraudNetConfig',
-				array(
-					'f'       => $this->fraud_net->session_id(),
-					's'       => $this->fraud_net->source_website_id(),
-					'sandbox' => $this->environment->current_environment_is( Environment::SANDBOX ),
-				)
-			);
-		}
 	}
 }
